visabud-flask-server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.8
click==8.1.7
cmake==3.27.2
coloredlogs==15.0.1
contourpy==1.1.0
cssselect2==0.7.0
cycler==0.11.0
dataclasses-json==0.5.14
emoji==2.8.0
exceptiongroup==1.1.3
faiss-cpu==1.7.4
fastapi==0.99.1
filelock==3.12.3
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
httptools==0.6.0
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.4
langchain==0.0.278
langsmith==0.0.31
lit==16.0.6
llvmlite==0.40.1
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==9.0.0.312
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.15.1
openai==0.27.9
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
overrides==7.4.0
packaging==23.1
Pillow==10.0.0
posthog==3.0.2
protobuf==4.24.2
pulsar-client==3.3.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pydub==0.25.1
pydyf==0.8.0
pyparsing==3.0.9
pyphen==0.14.0
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.20
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.3.3
tinycss2==1.2.1
tokenizers==0.13.3
torch==2.0.1
tqdm==4.66.1
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
unstructured==0.10.10
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
weasyprint==60.1
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
zopfli==0.2.3
